Passing methods to the user controller
