+++ src/main/java/com/alkemy/ong/controller/UserController.java
@@ -1,13 +1,18 @@
 package com.alkemy.ong.controller;
 
+import com.alkemy.ong.dto.UserDto;
+import com.alkemy.ong.security.mapper.UserMapper;
 import com.alkemy.ong.service.UserService;
 import com.fasterxml.jackson.databind.introspect.TypeResolutionContext;
+import java.util.List;
 import javax.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -15,13 +20,26 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/users")
 public class UserController {
-    
+
     @Autowired
     private UserService userService;
-    
+    @Autowired
+    private UserMapper userMapper;
+
     @DeleteMapping("/{id}")
     public ResponseEntity<TypeResolutionContext.Empty> delete(@PathVariable Long id) throws EntityNotFoundException {
         userService.delete(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    }
+
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @GetMapping()
+    public ResponseEntity<List<UserDto>> getAllUsersD() {
+        return new ResponseEntity<List<UserDto>>(userService.getAllUsers(), HttpStatus.OK);
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<UserDto> findUserById(@PathVariable Long id) {
+        return ResponseEntity.ok().body(userMapper.convertUserToDto(userService.findUserById(id).get()));
     }
 }
